New Slab
Wrote a new program to generate the slab. After noticing a number of issues with the current slab I was using.
Modified gro2lmp such that suggested is the simulation scheme I'm after.

Modified add_water to fix box offset

Modified lmp_sol to leave a reasonable gap between slab and water and to gracefully handle input files with tabs before the data as lammps prefers.