Extra analysis
Couple of modifications to chitin_build.plin an attempt to get the closer to the right structure. But it is still broken.

charge_dist.pl is an attempt to calculate the distance betwen the alpha-carbons of two groups of charged residues. Inital attempt was using perlmol modules but these don't support the reading in the reasidue type, thus needs to be rewritten